Properly encode the requests as JSON
Currently the requests JSON are just built with some formatting which
can lead to invalid JSON. To solve this extend the `req` function to
automatically encode the `data` argument as JSON when it is a dict.
It now also set the request `Content-Type` header to
`application/json`.

Then fix all requests to pass their dict to `req`.